refactor(order): update consumer and producer
