+++ order/internal/app/di.go
@@ -22,9 +22,9 @@ import (
 	"github.com/andredubov/rocket-factory/order/internal/converter/kafka/decoder"
 	"github.com/andredubov/rocket-factory/order/internal/repository/order/postgres"
 	"github.com/andredubov/rocket-factory/order/internal/service"
-	consumer "github.com/andredubov/rocket-factory/order/internal/service/consumer/order"
+	consumer "github.com/andredubov/rocket-factory/order/internal/service/consumer/order_assembled_consumer"
 	orders "github.com/andredubov/rocket-factory/order/internal/service/order"
-	"github.com/andredubov/rocket-factory/order/internal/service/producer"
+	producer "github.com/andredubov/rocket-factory/order/internal/service/producer/order_paid_producer"
 	"github.com/andredubov/rocket-factory/platform/pkg/closer"
 	wrappedKafka "github.com/andredubov/rocket-factory/platform/pkg/kafka"
 	wrappedKafkaConsumer "github.com/andredubov/rocket-factory/platform/pkg/kafka/consumer"
