+++ order/internal/service/consumer/order_assembled_consumer/consumer.go
@@ -1,4 +1,4 @@
-package order
+package order_assembled_consumer
 
 import (
 	"context"
