+++ order/internal/service/consumer/order_assembled_consumer/handler.go
@@ -1,4 +1,4 @@
-package order
+package order_assembled_consumer
 
 import (
 	"context"
