+++ order/internal/service/consumer/order_assembled_consumer/tests/consumer_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	kafkaConverterMocks "github.com/andredubov/rocket-factory/order/internal/converter/kafka/mocks"
-	"github.com/andredubov/rocket-factory/order/internal/service/consumer/order"
+	"github.com/andredubov/rocket-factory/order/internal/service/consumer/order_assembled_consumer"
 	repoMocks "github.com/andredubov/rocket-factory/order/internal/service/mocks"
 	"github.com/andredubov/rocket-factory/platform/pkg/kafka/mocks"
 )
@@ -18,7 +18,7 @@ func TestNewService(t *testing.T) {
 	mockRepo := repoMocks.NewOrdersRepository(t)
 
 	// Вызываем тестируемую функцию
-	service := order.NewService(mockConsumer, mockDecoder, mockRepo)
+	service := order_assembled_consumer.NewService(mockConsumer, mockDecoder, mockRepo)
 
 	// Проверяем, что сервис создан корректно
 	assert.NotNil(t, service)
