+++ order/internal/service/producer/order_paid_producer/producer.go
@@ -1,4 +1,4 @@
-package producer
+package order_paid_producer
 
 import (
 	"context"
